chore(lmkv): split oversized graph_ops files for the Server Size Guard
The glqk/qoxm additions tipped three mcp_bridge/graph_ops files over the
1500-line/50KiB budget (guard checks changed files, so inherited debt
counts too):

- query.rs: path + edges ops move to graph_ops/edges_op.rs (same
  impl-split pattern as snapshot.rs/insights.rs)
- tests/snapshot.rs: context/edge-category op tests move to
  tests/context.rs (they were never snapshot tests)
- snapshot.rs: cross-module co-change pressure builder moves to
  refactor.rs beside the rest of the refactor_candidates ranking math

No behavior change; moved tests pass (130 graph_ops tests green
locally, 5 hotspots_churn failures are the usual no-local-Postgres).
