+++ server/src/mcp_bridge/graph_ops/tests/mod.rs
@@ -1301,6 +1301,7 @@ async fn context_imports_bucket_for_file_references_pr_c1() {
 }
 
 mod complexity_refactor;
+mod context;
 mod crate_graph;
 mod flow;
 mod registry_bridge_coverage;
